Improved scaffolding and structure
