+++ src/constants.js
@@ -1,0 +1,3 @@
+const ApiKey = "AIzaSyBf1UOdCb4_NT4h_g4Wzz4taWIok5cpeCQ";
+
+exports.ApiKey = ApiKey
