+++ src/index.js
@@ -1,556 +1,22 @@
 import React, { useState, useEffect } from "react";
 import ReactDOM from "react-dom";
-import GoogleMapReact from "google-map-react";
-import "./styles.css";
-import { useFilters, useTable } from "react-table";
-import logo from "./milano-2.png";
-import { BsPencilSquare } from "react-icons/bs";
-import { BsInstagram } from "react-icons/bs";
-import { BsFacebook } from "react-icons/bs";
-import { BsWhatsapp } from "react-icons/bs";
-import { BsCaretDownSquareFill } from "react-icons/bs";
-import { PiMicrophoneStageFill } from "react-icons/pi";
+import "./assets/styles.css";
 
-const { GoogleSpreadsheet } = require("google-spreadsheet");
-//const FacebookGroup = "https://www.facebook.com/groups/standupcomedyitalia/";
+import TableAndMap from "./TableAndMap";
+import NavigationBar from "./NavigationBar"
+
 const GitHubURL = "https://github.com/morettimarco/open-mic-nights";
-const GoogleForm = "https://forms.gle/vDuLfQ7Bc9iKxT2o8";
+
 const SpreadsheetId = "1_X_znvg8kGbFMXoys011182T5ZTGONCsveY9uLEWsr8";
-const SpreadsheetURL =
-  "https://docs.google.com/spreadsheets/d/" + SpreadsheetId;
-const ApiKey = "AIzaSyBf1UOdCb4_NT4h_g4Wzz4taWIok5cpeCQ";
+const SpreadsheetURL ="https://docs.google.com/spreadsheets/d/" + SpreadsheetId;
 
 // This is a custom filter UI for selecting
 // a unique option from a list
-function SelectColumnFilter({
-  column: { filterValue, setFilter, preFilteredRows, id },
-}) {
-  // Calculate the options for filtering
-  // using the preFilteredRows
-  const options = React.useMemo(() => {
-    const options = new Set();
-    preFilteredRows.forEach((row) => {
-      options.add(row.values[id]);
-    });
-    return [...options.values()];
-  }, [id, preFilteredRows]);
 
-  // Render a multi-select box
-  return (
-    <select
-      className="select is-small"
-      value={filterValue}
-      onChange={(e) => {
-        setFilter(e.target.value || undefined);
-      }}
-    >
-      <option value="">All</option>
-      {options.map((option, i) => (
-        <option key={i} value={option}>
-          {option}
-        </option>
-      ))}
-    </select>
-  );
-}
 
-function SearchColumnFilter({ column: { filterValue, setFilter } }) {
-  return (
-    <input
-      className="input is-small"
-      value={filterValue || ""}
-      onChange={(e) => setFilter(e.target.value)}
-    />
-  );
-}
-
-function TableAndMap() {
-  const [data, setData] = useState({
-    headerValues: null,
-    rows: [],
-    isFetching: false,
-  });
-
-  useEffect(() => {
-    (async function () {
-      try {
-        setData({ ...data, isFetching: true });
-
-        const doc = new GoogleSpreadsheet(SpreadsheetId);
-        doc.useApiKey(ApiKey);
-        await doc.loadInfo();
-
-        const sheet = doc.sheetsByIndex[0];
-        const rows = await sheet.getRows();
-
-        setData({
-          headerValues: sheet.headerValues,
-          rows: rows,
-          isFetching: false,
-        });
-      } catch (e) {
-        console.error(e);
-        setData({ ...data, isFetching: false });
-      }
-    })();
-  }, []);
-
-  const rowsData = React.useMemo(() => {
-    if (!data.isFetching && data.headerValues != null) {
-      return data.rows.map((row) => {
-        return {
-          Address: row["Address"],
-          AudienceEntryFee: row["Audience Entry Fee"],
-          HowToBook: row["Contact / Book a Spot"],
-          Language: row["Language"],
-          Category: row["Event Category"],
-          SubCategory: row["Event Sub-Category"],
-          Status: row["Status"],
-          OrganizerName: row["Organizer Name"],
-          Description: row["Event Description"],
-          FacebookGroup: row["Facebook Group"],
-          FacebookPage: row["Facebook Page"],
-          WhatsApp: row["WhatsApp"],
-          Frequency: row["Frequency"],
-          Indoor: row["Indoor / Outdoor"],
-          Instagram: row["Instagram"],
-          Latitude: row["Latitude"],
-          Level: row["Comedian Level"],
-          Longitude: row["Longitude"],
-          Name: row["Name"],
-          PayToPlay: row["Pay to Play"],
-          RowNumber: row.rowNumber,
-          Time: row["Event Time"],
-          UpdateInfoFormLink: row["Update Info Form Link"],
-          Venue: row["Venue"],
-          WalkIn: row["Walk-in"],
-          Weekday: row["Weekday / Month"],
-          Website: row["Website"],
-          WheelchairAccess: row["Wheelchair Access"],
-        };
-      });
-    }
-    return [];
-  }, [data]);
-
-  const columns = React.useMemo(() => {
-    if (!data.isFetching && data.headerValues != null) {
-      return [
-        {
-          Header: "",
-          accessor: "UpdateInfoFormLink",
-          disableFilters: true,
-          Cell: ({ row }) => {
-            return (
-              <a href={row.original.UpdateInfoFormLink} target="_blank">
-                <BsPencilSquare />
-              </a>
-            );
-          },
-        },
-        {
-          Header: "",
-          accessor: "Instagram",
-          disableFilters: true,
-          Cell: ({ row }) => {
-            return (
-              <div>
-                <a href={row.original.Instagram}>
-                  <BsInstagram />
-                </a>
-                <a href={row.original.FacebookPage}>
-                  <BsFacebook />
-                </a>
-                <a href={row.original.WhatsApp}>
-                  <BsWhatsapp />
-                </a>
-              </div>
-            );
-          },
-        },
-        {
-          Header: "Name",
-          accessor: "Name",
-          disableFilters: true,
-          Filter: SearchColumnFilter,
-        },
-        {
-          Header: "Description",
-          accessor: "Description",
-          hideInitially: true,
-          Filter: SearchColumnFilter,
-        },
-        {
-          Header: "Category",
-          accessor: "Category",
-          hideInitially: true,
-          Filter: SearchColumnFilter,
-        },
-        {
-          Header: "Sub Category",
-          accessor: "SubCategory",
-          hideInitially: true,
-          Filter: SearchColumnFilter,
-        },
-        {
-          Header: "Status",
-          accessor: "Status",
-          Filter: SelectColumnFilter,
-          filter: "equals",
-        },
-        {
-          Header: "Organizer Name",
-          accessor: "OrganizerName",
-          hideInitially: true,
-          Filter: SearchColumnFilter,
-        },
-        // {
-        //   Header: "Walk In",
-        //   accessor: "WalkIn",
-        //   hideInitially: true,
-        //   Filter: SelectColumnFilter,
-        // },
-        // {
-        //   Header: "Pay to Play",
-        //   accessor: "PayToPlay",
-        //   hideInitially: true,
-        //   Filter: SelectColumnFilter,
-        // },
-        {
-          Header: "Audience Entry Fee",
-          accessor: "AudienceEntryFee",
-          hideInitially: true,
-          Filter: SelectColumnFilter,
-        },
-        {
-          Header: "Level",
-          accessor: "Level",
-          Filter: SelectColumnFilter,
-          filter: "equals",
-        },
-        {
-          Header: "Language",
-          accessor: "Language",
-          Filter: SelectColumnFilter,
-        },
-        {
-          Header: "Frequency",
-          accessor: "Frequency",
-          Filter: SelectColumnFilter,
-          //hideInitially: true,
-        },
-        {
-          Header: "Weekday",
-          accessor: "Weekday",
-          Filter: SelectColumnFilter,
-        },
-        {
-          Header: "Time",
-          accessor: "Time",
-          Filter: SearchColumnFilter,
-        },
-        {
-          Header: "Venue",
-          accessor: "Venue",
-          Filter: SearchColumnFilter,
-          disableFilters: true,
-        },
-        {
-          Header: "Website",
-          accessor: "Website",
-          hideInitially: true,
-          disableFilters: true,
-        },
-        {
-          Header: "Address",
-          accessor: "Address",
-          Filter: SearchColumnFilter,
-          disableFilters: true,
-        },
-        {
-          Header: "Wheelchair Access",
-          accessor: "WheelchairAccess",
-          hideInitially: true,
-          Filter: SelectColumnFilter,
-        },
-        {
-          Header: "How To Book",
-          accessor: "HowToBook",
-          Filter: SearchColumnFilter,
-          disableFilters: true,
-        },
-        {
-          Header: "Facebook Group",
-          accessor: "FacebookGroup",
-          hideInitially: true,
-          disableFilters: true,
-          Cell: ({ row }) => {
-            return (
-              <a href={row.original.FacebookGroup}>
-                {row.original.FacebookGroup}
-              </a>
-            );
-          },
-        },
-        // {
-        //   Header: "Indoor?",
-        //   accessor: "Indoor",
-        //   hideInitially: true,
-        //   Filter: SelectColumnFilter,
-        // },
-      ];
-    }
-    return [];
-  }, [data]);
-
-  const initialFilterSettings = React.useMemo(() => {
-    return [{ id: "BackOn", value: "Yes" }];
-  });
-
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    rows,
-    prepareRow,
-    allColumns,
-  } = useTable(
-    {
-      columns,
-      data: rowsData,
-      initialState: {
-        hiddenColumns: columns
-          .filter((c) => c.hideInitially)
-          .map((c) => c.accessor),
-        filters: initialFilterSettings,
-      },
-    },
-    useFilters
-  );
-
-  // line is zero-based
-  // line is the row number that you want to see into view after scroll
-
-  return (
-    <>
-      <div className="columns is-multiline">
-        <span className="map column is-12-mobile is-5-desktop">
-          <Map results={rows} />
-        </span>
-        <span className="table_wrapper column is-12-mobile is-7-desktop">
-          <div id="table-dropdown" className="dropdown">
-            <div className="dropdown-trigger">
-              <button
-                className="button"
-                onClick={() =>
-                  document
-                    .getElementById("table-dropdown")
-                    .classList.toggle("is-active")
-                }
-              >
-                Select columns 🔽
-              </button>
-            </div>
-            <div className="dropdown-menu" id="dropdown-menu" role="menu">
-              {allColumns.map((column) => (
-                <div key={column.id} className="dropdown-content">
-                  <label className="checkbox">
-                    <input type="checkbox" {...column.getToggleHiddenProps()} />
-                    {column.id}
-                  </label>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          <table className="table is-hoverable" {...getTableProps()}>
-            <thead>
-              {headerGroups.map((headerGroup) => (
-                <tr {...headerGroup.getHeaderGroupProps()}>
-                  {headerGroup.headers.map((column) => (
-                    <th {...column.getHeaderProps()}>
-                      {column.render("Header")}
-                      <div>
-                        {column.canFilter ? column.render("Filter") : null}
-                      </div>
-                    </th>
-                  ))}
-                </tr>
-              ))}
-            </thead>
-            <tbody {...getTableBodyProps()}>
-              {rows.map((row) => {
-                prepareRow(row);
-                return (
-                  <tr {...row.getRowProps()}>
-                    {row.cells.map((cell) => {
-                      return (
-                        <td {...cell.getCellProps()}>{cell.render("Cell")}</td>
-                      );
-                    })}
-                  </tr>
-                );
-              })}
-            </tbody>
-          </table>
-        </span>
-      </div>
-    </>
-  );
-}
-
-const InfoWindow = (props) => {
-  const { name, address,  weekday} = props;
-  const infoWindowStyle = {
-    position: "relative",
-    bottom: 50,
-    left: "-45px",
-    textAlign: "center",
-    width: 220,
-    backgroundColor: "white",
-    boxShadow: "0 2px 7px 1px rgba(0, 0, 0, 0.3)",
-    padding: 10,
-    fontSize: 14,
-    zIndex: 100,
-    borderRadius: "25px",
-  };
-
-  return (
-    <div style={infoWindowStyle}>
-      <div style={{ fontSize: 14 }}>{name} </div>
-      <div style={{ fontSize: 10 }}>{weekday} at {address}</div>
-    </div>
-  );
-};
-
-class MapMarker extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    const markerStyle = {
-      height: 25,
-      width: 25,
-      zIndex: 10,
-      cursor: "pointer",
-    };
-    return (
-      <div >
-        <PiMicrophoneStageFill style={markerStyle} className={
-          this.props.status === "Active"
-            ? "is-link"
-            : "is-danger"
-        }  />
-        {this.props.show && <InfoWindow name={this.props.name} address={this.props.address} weekday={this.props.weekday}  status={this.props.status}/>}
-      </div>
-    );
-  }
-}
-
-class Map extends React.Component {
-  // TODO: turn this into a function
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      results: [],
-    };
-  }
-
-  _onChildClick = (key, childProps) => {
-    this.setState((state) => {
-      let index = state.results.findIndex((e) => e.show);
-      if (
-        index > 0 &&
-        state.results[index].original.RowNumber != parseInt(key)
-      ) {
-        state.results[index].show = false;
-      }
-      index = state.results.findIndex(
-        (e) => e.original.RowNumber == parseInt(key)
-      );
-      state.results[index].show = !state.results[index].show; // eslint-disable-line no-param-reassign
-
-      const cells = document.querySelectorAll("td");
-      var found = null;
-      for (var i = 0; i < cells.length; i++) {
-        if (cells[i].textContent == state.results[index].original.Name) {
-          found = cells[i];
-          break;
-        }
-      }
-      found.parentNode.style.backgroundColor = "#fffcb3";
-      setTimeout(
-        () => (found.parentNode.style.backgroundColor = "white"),
-        2000
-      );
-      found.scrollIntoView({ behavior: "smooth", block: "nearest" });
-
-      return { results: state.results };
-    });
-  };
-
-  componentDidUpdate(prevProps) {
-    if (this.props.results != prevProps.results) {
-      this.setState((_, props) => {
-        const newState = props.results.map((result) => ({
-          ...result,
-          show: false,
-        }));
-        return { results: newState };
-      });
-    }
-  }
-
-  render() {
-    return (
-      // Important! Always set the container height explicitly
-      <div id="map">
-        <GoogleMapReact
-          bootstrapURLKeys={{ key: ApiKey }}
-          defaultCenter={{ lat: 45.463336, lng: 9.187174 }}
-          defaultZoom={12.8}
-          onChildClick={this._onChildClick}
-        >
-          {this.state.results.map((item) => (
-            <MapMarker
-              key={item.original.RowNumber}
-              lat={item.original.Latitude}
-              lng={item.original.Longitude}
-              address={item.original.Address}
-              weekday={item.original.Weekday}
-              name={item.original.Name}
-              status={item.original.Status}
-              show={item.show}
-            />
-          ))}
-        </GoogleMapReact>
-      </div>
-    );
-  }
-}
 
 function App() {
   const qna = [
-    //   {
-    //     question: "My night's details are out of date! ⌚",
-    //     answer: (
-    //       <p>
-    //         Just <a href="https://tripetto.app/run/ZNRADVJBA8">let us know</a>!
-    //         We'll fix it ASAP.
-    //       </p>
-    //     ),
-    //   },
-    // {
-    //   question: "I have feedback to share! 🤔",
-    //   answer: (
-    //     <p>
-    //       We're still in early development, so any feedback is useful for us -
-    //       send it <a href="https://tripetto.app/run/ZNRADVJBA8">here</a>.
-    //     </p>
-    //   ),
-    // },
     {
       question: "Found a bug? Wanna contribute? Rip the site and f**k us?",
       answer: (
@@ -584,47 +50,7 @@ function App() {
   ];
   return (
     <>
-      <nav className="navbar is-light has-shadow py-4 mb-2">
-        <div className="navbar-brand">
-          <a className="navbar-item">
-            <img src={logo} style={{ maxHeight: "150px" }} />
-          </a>
-          <div className="navbar-start">
-            <div className="navbar-item">
-              <div className="title-subtitle-container">
-                <p className="title">Milan Standup Comedy Map</p>
-                <p className="subtitle">Perform comedy near you</p>
-              </div>
-            </div>
-            <div
-              className="navbar-burger"
-              onClick={() =>
-                document
-                  .getElementById("nav-links")
-                  .classList.toggle("is-active")
-              }
-            >
-              <BsCaretDownSquareFill fontSize="3em" />
-            </div>
-          </div>
-        </div>
-        <div className="navbar-menu" id="nav-links">
-          <div className="navbar-end">
-            <a className="navbar-item navbar-item-centered" href={GoogleForm}>
-              🎤 Submit an open mic night
-            </a>
-            <a className="js-modal-trigger navbar-item-centered" data-target="modal-js-example">
-            ❓ F.A.Q.
-            </a>
-            <a
-              className="navbar-item navbar-item-centered"
-              href="https://www.instagram.com/_anarchytect/"
-            >
-              📣 Contact me for feedbacks!
-            </a>
-          </div>
-        </div>
-      </nav>
+      <NavigationBar/>
 
       <TableAndMap />
 
@@ -646,64 +72,11 @@ function App() {
               );
             })}
           </div>
-          {/* <p className="has-text-centered my-2 mx-2">
-            Website created with ❤ by{" "}
-            <a href="https://apuchitnis.github.io/">@apuchitnis</a>. Thanks to{" "}
-            <a href="https://www.facebook.com/GaelleConstantComedian">
-              Gaelle Constant
-            </a>{" "}
-            for maintaining the{" "}
-            <a href={SpreadsheetURL}>
-              original spreadsheet
-            </a>
-            .
-          </p> */}
         </div>
       </div>
     </>
   );
 }
-document.addEventListener('DOMContentLoaded', () => {
-  // Functions to open and close a modal
-  function openModal($el) {
-    $el.classList.add('is-active');
-  }
 
-  function closeModal($el) {
-    $el.classList.remove('is-active');
-  }
-
-  function closeAllModals() {
-    (document.querySelectorAll('.modal') || []).forEach(($modal) => {
-      closeModal($modal);
-    });
-  }
-
-  // Add a click event on buttons to open a specific modal
-  (document.querySelectorAll('.js-modal-trigger') || []).forEach(($trigger) => {
-    const modal = $trigger.dataset.target;
-    const $target = document.getElementById(modal);
-
-    $trigger.addEventListener('click', () => {
-      openModal($target);
-    });
-  });
-
-  // Add a click event on various child elements to close the parent modal
-  (document.querySelectorAll('.modal-background, .modal-close, .modal-card-head .delete, .modal-card-foot .button') || []).forEach(($close) => {
-    const $target = $close.closest('.modal');
-
-    $close.addEventListener('click', () => {
-      closeModal($target);
-    });
-  });
-
-  // Add a keyboard event to close all modals
-  document.addEventListener('keydown', (event) => {
-    if (event.code === 'Escape') {
-      closeAllModals();
-    }
-  });
-});
 const rootElement = document.getElementById("app");
 ReactDOM.render(<App />, rootElement);
